delete-wrong-pages
pagesが二つあったので削除しました。動作確認済みです。
