--- pages/index.tsx
@@ -1,6 +1,0 @@
-import type { FC } from 'react';
-
-const Entrance: FC = () => <div>Welcome to Toho Festival 2021!</div>;
-
-export default Entrance;
-
